requirements.txt
# 核心依赖
pyzmq>=25.0.0

# 数据序列化和验证
protobuf>=4.21.0
pydantic>=2.0.0

# 类型检查
typeguard>=4.0.0

# 日志和配置
pyyaml>=6.0

# 开发依赖（可选）
# pytest>=7.0.0
# pytest-cov>=4.0.0
# sphinx>=5.0.0
# sphinx-rtd-theme>=1.0.0